fix(audit): stop the ledger reporting itself as tampered on table overflow
AuditLedgerService caps its sequence table at 50,000 conversations and
cleared the whole thing on overflow, on the reasoning that re-seeding from
countByConversation was "correct, only slower". It is not: entries sit in
the queue for up to a flush interval, so a conversation with queued entries
has consumed positions the store cannot see yet, and the re-seed hands the
same position out twice. The verifier grades a duplicate exactly like a gap
(BROKEN), and unlike a gap there is no exculpatory record for it — so the
ledger accused the deployment of tampering because its own table wrapped
around. On a busy deployment the queue is never empty, so essentially every
overflow produced duplicates.

Evict only counters whose positions are all accounted for where a re-seed
can see them (persisted, or attributed in `undelivered`); retain the rest.
A ReentrantReadWriteLock spans "position consumed" to "entry visible in the
queue" against eviction, since that window contains HMAC and Ed25519
signing. flush() publishes its in-flight batch — between poll and a
successful append those positions are in neither queue nor store — and is
synchronized so the scheduled writer and the @PreDestroy flush cannot poll
interleaved halves. Still full after eviction yields UNSEQUENCED, which
degrades the window to UNAVAILABLE rather than to an accusation.

Both regression tests are mutation-checked: with the retain set emptied the
duplicate test fails with expected <[0, 1]> but was <[0, 0]>.

Also from the same review:

- AgentGroupStore rejected HUMAN_DECIDES as needing "human group members
  (I6), which are not available yet" — 150 lines below its own I6 matrix
  that validates those members. They shipped in 10c; the missing piece is
  the resume path a paused tie-break needs. The test pinned the word "I6",
  so it now asserts on the offered alternatives instead.
- SafeHttpClient claimed an overall wall-clock timeout across all hops, but
  the budget is only checked between hops, so a hop that trickles its body
  hung forever. Bound both the initial and redirect hops by one default and
  state what is actually true.
- ToolRateLimiter.refill() overflowed `elapsedNanos * limit` after ~107 idle
  days, driving tokens negative so the bucket refused every later call.
- ConversationStepRunner registered the in-flight conversation outside the
  try whose finally unregisters it.

+++ src/main/java/ai/labs/eddi/modules/llm/tools/ToolRateLimiter.java
@@ -188,7 +188,12 @@ public class ToolRateLimiter {
                 return;
             }
             lastRefillNanos = now;
-            tokens = Math.min(limit, tokens + elapsedNanos * limit / WINDOW_NANOS);
+            // (double) on elapsedNanos: `elapsedNanos * limit` is long arithmetic and
+            // overflows once a bucket has been idle for ~107 days at the default
+            // limit of 1000. The wrapped negative would drive `tokens` below zero and
+            // tryAcquire would refuse every call from then on — a bucket that silently
+            // locks shut rather than refilling.
+            tokens = Math.min(limit, tokens + (double) elapsedNanos * limit / WINDOW_NANOS);
         }
 
         synchronized boolean tryAcquire() {
